user_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyMySQL==1.1.0
cryptography==42.0.8
gunicorn==21.2.0
django-cors-headers==4.3.0
